Refactor JSON construction/conversion into its own service, with tests
