+++ restapi/src/test/java/io/stargate/web/docsapi/service/DocumentServiceTest.java
@@ -22,9 +22,7 @@ import com.datastax.oss.driver.api.core.ProtocolVersion;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.BooleanNode;
 import com.fasterxml.jackson.databind.node.IntNode;
-import com.fasterxml.jackson.databind.node.NullNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.common.collect.ImmutableList;
 import com.google.gson.JsonArray;
@@ -58,7 +56,6 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.ExecutionException;
 import javax.ws.rs.core.MultivaluedMap;
@@ -999,8 +996,7 @@ public class DocumentServiceTest {
   }
 
   @Test
-  public void getJsonAtPath()
-      throws ExecutionException, InterruptedException, UnauthorizedException {
+  public void getJsonAtPath() throws UnauthorizedException {
     AuthorizationService authorizationService = mock(AuthorizationService.class);
     DocumentDB dbMock = mock(DocumentDB.class);
     ResultSet rsMock = mock(ResultSet.class);
@@ -1020,12 +1016,14 @@ public class DocumentServiceTest {
   }
 
   @Test
-  public void getJsonAtPath_withRowsEmptyJson()
-      throws ExecutionException, InterruptedException, UnauthorizedException {
+  public void getJsonAtPath_withRowsEmptyJson() throws UnauthorizedException {
     AuthorizationService authorizationService = mock(AuthorizationService.class);
     DocumentDB dbMock = mock(DocumentDB.class);
     ResultSet rsMock = mock(ResultSet.class);
     DocumentService serviceMock = mock(DocumentService.class, CALLS_REAL_METHODS);
+    JsonConverterService jsonConverterServiceMock =
+        mock(JsonConverterService.class, CALLS_REAL_METHODS);
+    when(serviceMock.getJsonConverterService()).thenReturn(jsonConverterServiceMock);
     when(dbMock.executeSelect(anyString(), anyString(), anyListOf(BuiltCondition.class)))
         .thenReturn(rsMock);
     when(dbMock.getAuthorizationService()).thenReturn(authorizationService);
@@ -1039,7 +1037,8 @@ public class DocumentServiceTest {
 
     List<PathSegment> path = smallPath();
 
-    when(serviceMock.convertToJsonDoc(anyListOf(Row.class), anyBoolean(), anyBoolean()))
+    when(jsonConverterServiceMock.convertToJsonDoc(
+            anyListOf(Row.class), anyBoolean(), anyBoolean()))
         .thenReturn(ImmutablePair.of(mapper.createObjectNode(), new HashMap<>()));
 
     JsonNode result = serviceMock.getJsonAtPath(dbMock, "ks", "collection", "id", path);
@@ -1048,13 +1047,14 @@ public class DocumentServiceTest {
   }
 
   @Test
-  public void getJsonAtPath_withRows()
-      throws ExecutionException, InterruptedException, JsonProcessingException,
-          UnauthorizedException {
+  public void getJsonAtPath_withRows() throws JsonProcessingException, UnauthorizedException {
     AuthorizationService authorizationService = mock(AuthorizationService.class);
     DocumentDB dbMock = mock(DocumentDB.class);
     ResultSet rsMock = mock(ResultSet.class);
     DocumentService serviceMock = mock(DocumentService.class, CALLS_REAL_METHODS);
+    JsonConverterService jsonConverterServiceMock =
+        mock(JsonConverterService.class, CALLS_REAL_METHODS);
+    when(serviceMock.getJsonConverterService()).thenReturn(jsonConverterServiceMock);
     when(dbMock.executeSelect(anyString(), anyString(), anyListOf(BuiltCondition.class)))
         .thenReturn(rsMock);
     when(dbMock.getAuthorizationService()).thenReturn(authorizationService);
@@ -1071,7 +1071,8 @@ public class DocumentServiceTest {
     ObjectNode jsonObj = mapper.createObjectNode();
     jsonObj.set("abc", mapper.readTree("[1]"));
 
-    when(serviceMock.convertToJsonDoc(anyListOf(Row.class), anyBoolean(), anyBoolean()))
+    when(jsonConverterServiceMock.convertToJsonDoc(
+            anyListOf(Row.class), anyBoolean(), anyBoolean()))
         .thenReturn(ImmutablePair.of(jsonObj, new HashMap<>()));
 
     JsonNode result = serviceMock.getJsonAtPath(dbMock, "ks", "collection", "id", path);
@@ -1087,6 +1088,9 @@ public class DocumentServiceTest {
     DocumentDB dbMock = mock(DocumentDB.class);
     ResultSet rsMock = mock(ResultSet.class);
     DocumentService serviceMock = mock(DocumentService.class, CALLS_REAL_METHODS);
+    JsonConverterService jsonConverterServiceMock =
+        mock(JsonConverterService.class, CALLS_REAL_METHODS);
+    when(serviceMock.getJsonConverterService()).thenReturn(jsonConverterServiceMock);
     when(dbMock.executeSelect(anyString(), anyString(), anyListOf(BuiltCondition.class)))
         .thenReturn(rsMock);
     when(dbMock.getAuthorizationService()).thenReturn(authorizationService);
@@ -1105,7 +1109,8 @@ public class DocumentServiceTest {
 
     Map<String, List<JsonNode>> deadLeaves = new HashMap<>();
     deadLeaves.put("a", new ArrayList<>());
-    when(serviceMock.convertToJsonDoc(anyListOf(Row.class), anyBoolean(), anyBoolean()))
+    when(jsonConverterServiceMock.convertToJsonDoc(
+            anyListOf(Row.class), anyBoolean(), anyBoolean()))
         .thenReturn(ImmutablePair.of(jsonObj, deadLeaves));
 
     JsonNode result = serviceMock.getJsonAtPath(dbMock, "ks", "collection", "id", path);
@@ -1233,8 +1238,11 @@ public class DocumentServiceTest {
   public void searchDocumentsV2_emptyResult() throws Exception {
     DocumentDB dbMock = Mockito.mock(DocumentDB.class);
     DocumentService serviceMock = Mockito.mock(DocumentService.class);
+    JsonConverterService jsonConverterServiceMock =
+        mock(JsonConverterService.class, CALLS_REAL_METHODS);
     Mockito.when(serviceMock.searchDocumentsV2(any(), any(), any(), any(), any(), any(), any()))
         .thenCallRealMethod();
+    Mockito.when(serviceMock.getJsonConverterService()).thenReturn(jsonConverterServiceMock);
     Mockito.when(
             serviceMock.searchRows(
                 any(), any(), any(), any(), any(), any(), any(), any(), any(), any()))
@@ -1255,13 +1263,16 @@ public class DocumentServiceTest {
   public void searchDocumentsV2_existingResult() throws Exception {
     DocumentDB dbMock = Mockito.mock(DocumentDB.class);
     DocumentService serviceMock = Mockito.mock(DocumentService.class);
+    JsonConverterService jsonConverterServiceMock =
+        mock(JsonConverterService.class, CALLS_REAL_METHODS);
     Mockito.when(serviceMock.searchDocumentsV2(any(), any(), any(), any(), any(), any(), any()))
         .thenCallRealMethod();
+    Mockito.when(serviceMock.getJsonConverterService()).thenReturn(jsonConverterServiceMock);
     Mockito.when(
             serviceMock.searchRows(
                 any(), any(), any(), any(), any(), any(), any(), any(), any(), any()))
         .thenReturn(makeInitialRowData());
-    Mockito.when(serviceMock.convertToJsonDoc(any(), anyBoolean(), anyBoolean()))
+    Mockito.when(jsonConverterServiceMock.convertToJsonDoc(anyList(), anyBoolean(), anyBoolean()))
         .thenReturn(ImmutablePair.of(mapper.readTree("{\"a\": 1}"), new HashMap<>()));
     int pageSizeParam = 0;
     Paginator paginator = new Paginator(null, pageSizeParam, DocumentDB.SEARCH_PAGE_SIZE);
@@ -1280,13 +1291,16 @@ public class DocumentServiceTest {
   public void searchDocumentsV2_existingResultWithFields() throws Exception {
     DocumentDB dbMock = Mockito.mock(DocumentDB.class);
     DocumentService serviceMock = Mockito.mock(DocumentService.class);
+    JsonConverterService jsonConverterServiceMock = Mockito.mock(JsonConverterService.class);
+    Mockito.when(jsonConverterServiceMock.getMapper()).thenCallRealMethod();
     Mockito.when(serviceMock.searchDocumentsV2(any(), any(), any(), any(), any(), any(), any()))
         .thenCallRealMethod();
     Mockito.when(
             serviceMock.searchRows(
                 any(), any(), any(), any(), any(), any(), any(), any(), any(), any()))
         .thenReturn(makeInitialRowData());
-    Mockito.when(serviceMock.convertToJsonDoc(any(), anyBoolean(), anyBoolean()))
+    Mockito.when(serviceMock.getJsonConverterService()).thenReturn(jsonConverterServiceMock);
+    Mockito.when(jsonConverterServiceMock.convertToJsonDoc(any(), anyBoolean(), anyBoolean()))
         .thenReturn(ImmutablePair.of(mapper.readTree("{\"a\": 1}"), new HashMap<>()));
     List<FilterCondition> filters =
         ImmutableList.of(
@@ -1321,6 +1335,8 @@ public class DocumentServiceTest {
     Db dbFactoryMock = Mockito.mock(Db.class);
     DocumentDB dbMock = Mockito.mock(DocumentDB.class);
     DocumentService serviceMock = Mockito.mock(DocumentService.class);
+    JsonConverterService jsonConverterServiceMock =
+        mock(JsonConverterService.class, CALLS_REAL_METHODS);
     Mockito.when(dbFactoryMock.getDocDataStoreForToken(anyString(), any())).thenReturn(dbMock);
     Mockito.when(
             serviceMock.searchRows(
@@ -1331,8 +1347,9 @@ public class DocumentServiceTest {
                 any(), anyString(), anyString(), anyListOf(String.class), any()))
         .thenCallRealMethod();
     Mockito.doCallRealMethod().when(serviceMock).addRowsToMap(anyMap(), anyList());
-    Mockito.when(serviceMock.convertToJsonDoc(any(), anyBoolean(), anyBoolean()))
+    Mockito.when(jsonConverterServiceMock.convertToJsonDoc(anyList(), anyBoolean(), anyBoolean()))
         .thenReturn(ImmutablePair.of(mapper.readTree("{\"a\": 1}"), new HashMap<>()));
+    Mockito.when(serviceMock.getJsonConverterService()).thenReturn(jsonConverterServiceMock);
     int pageSizeParam = 0;
     Paginator paginator = new Paginator(null, pageSizeParam, DocumentDB.SEARCH_PAGE_SIZE);
 
@@ -1348,9 +1365,12 @@ public class DocumentServiceTest {
     Db dbFactoryMock = Mockito.mock(Db.class);
     DocumentDB dbMock = Mockito.mock(DocumentDB.class);
     DocumentService serviceMock = Mockito.mock(DocumentService.class);
+    JsonConverterService jsonConverterServiceMock =
+        mock(JsonConverterService.class, CALLS_REAL_METHODS);
     List<Row> twoDocsRows = makeInitialRowData();
     twoDocsRows.addAll(makeRowDataForSecondDoc());
     Mockito.when(dbFactoryMock.getDocDataStoreForToken(anyString(), any())).thenReturn(dbMock);
+    Mockito.when(serviceMock.getJsonConverterService()).thenReturn(jsonConverterServiceMock);
     Mockito.when(
             serviceMock.searchRows(
                 any(), any(), any(), any(), any(), any(), any(), any(), any(), any()))
@@ -1360,8 +1380,9 @@ public class DocumentServiceTest {
                 any(), anyString(), anyString(), anyListOf(String.class), any()))
         .thenCallRealMethod();
     Mockito.doCallRealMethod().when(serviceMock).addRowsToMap(anyMap(), anyList());
-    Mockito.when(serviceMock.convertToJsonDoc(any(), anyBoolean(), anyBoolean()))
+    Mockito.when(jsonConverterServiceMock.convertToJsonDoc(anyList(), anyBoolean(), anyBoolean()))
         .thenReturn(ImmutablePair.of(mapper.readTree("{\"a\": 1}"), new HashMap<>()));
+    when(serviceMock.getJsonConverterService()).thenReturn(jsonConverterServiceMock);
     int pageSizeParam = 0;
     Paginator paginator = new Paginator(null, pageSizeParam, DocumentDB.SEARCH_PAGE_SIZE);
 
@@ -1730,103 +1751,7 @@ public class DocumentServiceTest {
     assertThat(res).isNull();
   }
 
-  @Test
-  public void convertToJsonDoc_testDeadLeaves() throws JsonProcessingException {
-    List<Row> initial = makeInitialRowData();
-    initial.sort(
-        (row1, row2) ->
-            (Objects.requireNonNull(row1.getString("p0"))
-                        .compareTo(Objects.requireNonNull(row2.getString("p0")))
-                    * 100000
-                + Objects.requireNonNull(row1.getString("p1"))
-                        .compareTo(Objects.requireNonNull(row2.getString("p1")))
-                    * 10000
-                + Objects.requireNonNull(row1.getString("p2"))
-                        .compareTo(Objects.requireNonNull(row2.getString("p2")))
-                    * 1000
-                + Objects.requireNonNull(row1.getString("p3"))
-                        .compareTo(Objects.requireNonNull(row2.getString("p3")))
-                    * 100));
-    ImmutablePair<JsonNode, Map<String, List<JsonNode>>> result =
-        service.convertToJsonDoc(initial, false, false);
-
-    assertThat(result.left.toString())
-        .isEqualTo(
-            mapper
-                .readTree("{\"a\": {\"b\": {\"c\": true}}, \"d\": {\"e\": [3]}, \"f\": \"abc\"}")
-                .toString());
-
-    // This state should have no dead leaves, as it's the initial write
-    assertThat(result.right).isEmpty();
-
-    initial.addAll(makeSecondRowData());
-    initial.sort(
-        (row1, row2) ->
-            (Objects.requireNonNull(row1.getString("p0"))
-                        .compareTo(Objects.requireNonNull(row2.getString("p0")))
-                    * 100000
-                + Objects.requireNonNull(row1.getString("p1"))
-                        .compareTo(Objects.requireNonNull(row2.getString("p1")))
-                    * 10000
-                + Objects.requireNonNull(row1.getString("p2"))
-                        .compareTo(Objects.requireNonNull(row2.getString("p2")))
-                    * 1000
-                + Objects.requireNonNull(row1.getString("p3"))
-                        .compareTo(Objects.requireNonNull(row2.getString("p3")))
-                    * 100));
-    result = service.convertToJsonDoc(initial, false, false);
-
-    assertThat(result.left.toString())
-        .isEqualTo(
-            mapper
-                .readTree(
-                    "{\"a\": {\"b\": {\"c\": {\"d\": \"replaced\"}}}, \"d\": {\"e\": [3]}, \"f\": \"abc\"}")
-                .toString());
-
-    // This state should have 1 dead leaf, since $.a.b.c was changed from `true` to an object
-    Map<String, List<JsonNode>> expected = new HashMap<>();
-    List<JsonNode> list = new ArrayList<>();
-    ObjectNode node = mapper.createObjectNode();
-    node.set("", BooleanNode.valueOf(true));
-    list.add(node);
-    expected.put("$.a.b.c", list);
-    assertThat(result.right).isEqualTo(expected);
-
-    initial.addAll(makeThirdRowData());
-    initial.sort(
-        (row1, row2) ->
-            (Objects.requireNonNull(row1.getString("p0"))
-                        .compareTo(Objects.requireNonNull(row2.getString("p0")))
-                    * 100000
-                + Objects.requireNonNull(row1.getString("p1"))
-                        .compareTo(Objects.requireNonNull(row2.getString("p1")))
-                    * 10000
-                + Objects.requireNonNull(row1.getString("p2"))
-                        .compareTo(Objects.requireNonNull(row2.getString("p2")))
-                    * 1000
-                + Objects.requireNonNull(row1.getString("p3"))
-                        .compareTo(Objects.requireNonNull(row2.getString("p3")))
-                    * 100));
-    result = service.convertToJsonDoc(initial, false, false);
-
-    assertThat(result.left.toString()).isEqualTo(mapper.readTree("[\"replaced\"]").toString());
-
-    // This state should have 3 dead branches representing keys a, d, and f, since everything was
-    // blown away by the latest change
-    expected = new HashMap<>();
-    list = new ArrayList<>();
-    node = mapper.createObjectNode();
-    node.set("a", NullNode.getInstance());
-    node.set("d", NullNode.getInstance());
-    node.set("f", NullNode.getInstance());
-
-    list.add(node);
-    expected.put("$", list);
-
-    assertThat(result.right).isEqualTo(expected);
-  }
-
-  private List<Row> makeInitialRowData() {
+  public static List<Row> makeInitialRowData() {
     List<Row> rows = new ArrayList<>();
     Map<String, Object> data1 = new HashMap<>();
     Map<String, Object> data2 = new HashMap<>();
@@ -1867,7 +1792,7 @@ public class DocumentServiceTest {
     return rows;
   }
 
-  private List<Row> makeSecondRowData() {
+  public static List<Row> makeSecondRowData() {
     List<Row> rows = new ArrayList<>();
     Map<String, Object> data1 = new HashMap<>();
     data1.put("key", "1");
@@ -1883,7 +1808,7 @@ public class DocumentServiceTest {
     return rows;
   }
 
-  private List<Row> makeThirdRowData() {
+  public static List<Row> makeThirdRowData() {
     List<Row> rows = new ArrayList<>();
     Map<String, Object> data1 = new HashMap<>();
 
@@ -1900,7 +1825,7 @@ public class DocumentServiceTest {
     return rows;
   }
 
-  private List<Row> makeRowDataForSecondDoc() {
+  public List<Row> makeRowDataForSecondDoc() {
     List<Row> rows = new ArrayList<>();
     Map<String, Object> data1 = new HashMap<>();
